Update contracts and tests

+++ contracts/lib/SkrybeLib.sol
@@ -3,8 +3,8 @@ pragma solidity ^0.8.21;
 
 struct CollectionParams {
     string collectionId;
-    uint32 launchTimestamp;
-    uint32 whitelistLaunchTimestamp;
+    uint64 launchTimestamp;
+    uint64 whitelistLaunchTimestamp;
     uint256 price;
     uint256 whitelistPrice;
     uint32 maxSupply;
